chore: separate models and conform to Identifiable
