+++ MusicApp/Models/FollowersObject.swift
@@ -1,0 +1,13 @@
+//
+//  FollowersObject.swift
+//  MusicApp
+//
+//  Created by Miguel Sandoval on 31/07/24.
+//
+
+import Foundation
+
+struct FollowersObject: Codable {
+    let href: String?
+    let total: Int
+}
