+++ MusicApp/Models/ImageObject.swift
@@ -1,0 +1,14 @@
+//
+//  ImageObject.swift
+//  MusicApp
+//
+//  Created by Miguel Sandoval on 31/07/24.
+//
+
+import Foundation
+
+struct ImageObject: Codable {
+    let height: Int
+    let url: String
+    let width: Int
+}
