+++ MusicApp/Views/Home/ArtistList.swift
@@ -1,0 +1,20 @@
+//
+//  ArtistListView.swift
+//  MusicApp
+//
+//  Created by Miguel Sandoval on 31/07/24.
+//
+
+import SwiftUI
+
+struct ArtistListView: View {
+    var body: some View {
+        VStack {
+            
+        }
+    }
+}
+
+#Preview {
+    ArtistListView()
+}
