+++ MusicApp/Views/Utils/GenreTag.swift
@@ -1,0 +1,18 @@
+//
+//  GenrePill.swift
+//  MusicApp
+//
+//  Created by Miguel Sandoval on 31/07/24.
+//
+
+import SwiftUI
+
+struct GenrePill: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    GenrePill()
+}
