+++ MusicApp/Views/Utils/GradientBackground.swift
@@ -1,0 +1,18 @@
+//
+//  GradientBackground.swift
+//  MusicApp
+//
+//  Created by Miguel Sandoval on 31/07/24.
+//
+
+import SwiftUI
+
+struct GradientBackground: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    GradientBackground()
+}
